requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1